Add some missing events for the resolution, support loser payoff in the payoff path

+++ src/maths.rs
@@ -57,7 +57,7 @@ pub fn ninetails_payoff_losers(
     user_boosted_shares: U256,
     global_boosted_shares: U256,
 ) -> Result<U256, Error> {
-    let refund = mul_div(leftovers, U256::from(3e6), U256::from(10e6))?.0;
+    let refund = mul_div(leftovers, U256::from(3e6 as u32), U256::from(10e6 as u32))?.0;
     Ok(mul_div(user_boosted_shares, refund, global_boosted_shares)?.0)
 }
 
@@ -70,7 +70,7 @@ pub fn ninetails_payoff_winners(
     outcome_boosted_shares: U256,
     global_boosted_shares: U256,
 ) -> Result<U256, Error> {
-    let boost = mul_div(leftovers, U256::from(7e6), U256::from(10e6))?.0;
+    let boost = mul_div(leftovers, U256::from(7e6 as u32), U256::from(10e6 as u32))?.0;
     let winnings = mul_div(user_boosted_shares, boost, outcome_boosted_shares)?.0;
     let refund = ninetails_payoff_losers(leftovers, user_boosted_shares, global_boosted_shares)?;
     Ok(add!(winnings, refund))
